# apps/api/interstellar_api/routers/m2_calculations.py
"""M2 synchronous astronomical calculation endpoint."""

from __future__ import annotations

from datetime import UTC, datetime
from typing import Any, Literal
from uuid import uuid4

from fastapi import APIRouter, Query, Request, Response, status
from interstellar_core.application.astronomical_snapshot import (
    AstronomicalSnapshotInputError,
    create_astronomical_snapshot,
)
from interstellar_core.application.natal_technical_export import (
    NatalTechnicalExportError,
    render_natal_technical_document,
)
from interstellar_core.application.snapshot_tables import (
    SnapshotTableError,
    build_snapshot_table,
    table_json_bytes,
)
from interstellar_core.astronomy.adapters import SwissEphemerisAdapter
from pydantic import BaseModel, ConfigDict, Field, model_validator

from interstellar_api.errors import ErrorCode, ProblemException
from interstellar_api.workflow_store import WorkflowRecordNotFound

router = APIRouter(prefix="/api/v1", tags=["M2 Calculations"])


class StrictModel(BaseModel):
    model_config = ConfigDict(extra="forbid")


class SubjectReferencePayload(StrictModel):
    subject_version_id: str | None = Field(default=None, min_length=1, max_length=160)
    inline_subject: dict[str, Any] | None = None

    @model_validator(mode="after")
    def exactly_one_subject(self) -> SubjectReferencePayload:
        if (self.subject_version_id is None) == (self.inline_subject is None):
            raise ValueError("provide exactly one of subject_version_id or inline_subject")
        return self


class ChartDefinitionPayload(StrictModel):
    family: Literal[
        "natal",
        "transit",
        "progression",
        "direction",
        "return",
        "relationship",
        "horary",
        "electional",
        "relocation",
        "mundane",
        "harmonic",
        "custom",
    ]
    technique: str = Field(min_length=1, max_length=160)
    reference_time: dict[str, Any] | None = None
    reference_location: dict[str, Any] | None = None
    comparison_subjects: list[SubjectReferencePayload] = Field(default_factory=list)


class OrbOverridePayload(StrictModel):
    scope: Literal["point_pair", "point_class", "aspect", "chart_context"]
    point_a: str | None = None
    point_b: str | None = None
    point_class: str | None = None
    aspect_id: str | None = None
    chart_context: str | None = None
    orb_deg: float = Field(ge=0, le=30)


class ClassicalSettingsPayload(StrictModel):
    rulership_system: str | None = None
    dignity_table: str | None = None
    triplicity_table: str | None = None
    terms_table: str | None = None
    decan_or_face_table: str | None = None
    sect_rules: str | None = None
    lot_formula_set: str | None = None


class ChartSettingsPayload(StrictModel):
    calculation_profile_id: str | None = Field(default=None, min_length=1, max_length=160)
    analysis_system_id: str | None = Field(default=None, min_length=1, max_length=160)
    zodiac: Literal["tropical", "sidereal", "draconic"]
    ayanamsa: str | None = None
    house_system: str = Field(min_length=1, max_length=40)
    center: Literal["geocentric", "heliocentric", "topocentric"]
    coordinate_frame: Literal["ecliptic", "equatorial", "horizontal"] | None = None
    node_type: Literal["true", "mean", "both"]
    high_latitude_policy: Literal[
        "block", "allow_distorted", "fallback_whole_sign", "fallback_equal"
    ] | None = None
    aspect_set_id: str = Field(min_length=1, max_length=160)
    orb_profile_id: str = Field(min_length=1, max_length=160)
    point_set_ids: list[str] = Field(default_factory=list)
    included_points: list[str] = Field(default_factory=list)
    minor_body_ids: list[str] = Field(default_factory=list)
    fixed_star_ids: list[str] = Field(default_factory=list)
    lot_formula_ids: list[str] = Field(default_factory=list)
    hypothetical_point_ids: list[str] = Field(default_factory=list)
    included_aspect_ids: list[str] = Field(default_factory=list)
    orb_overrides: list[OrbOverridePayload] = Field(default_factory=list)
    classical_settings: ClassicalSettingsPayload | None = None
    custom_parameters: dict[str, Any] = Field(default_factory=dict)


class VersionReferencePayload(StrictModel):
    id: str = Field(min_length=1, max_length=160)
    version: str = Field(min_length=1)
    content_hash: str


class ChartRequestPayload(StrictModel):
    subject: SubjectReferencePayload
    chart: ChartDefinitionPayload
    settings: ChartSettingsPayload
    analysis_model: VersionReferencePayload | None = None
    rule_pack_hash: str = Field(pattern=r"^(?:sha256|hmac-sha256):[A-Fa-f0-9]{32,128}$")
    dataset_versions: dict[str, str]
    outputs: list[
        Literal[
            "snapshot",
            "default_render_manifest",
            "svg",
            "png",
            "pdf",
            "json",
            "csv",
            "ics",
            "markdown_technical",
            "plaintext_technical",
        ]
    ] = Field(min_length=1)
    input_fingerprint: str | None = None


def _unsupported(fields: dict[str, Any]) -> ProblemException:
    return ProblemException(
        status=status.HTTP_422_UNPROCESSABLE_CONTENT,
        code=ErrorCode.INVALID_REQUEST,
        detail="The M2 astronomical slice cannot execute the requested settings without guessing.",
        fields=fields,
    )


@router.post("/calculations", status_code=status.HTTP_201_CREATED)
async def create_calculation(payload: ChartRequestPayload, request: Request) -> dict[str, Any]:
    if payload.subject.inline_subject is not None:
        raise _unsupported(
            {"subject.inline_subject": "planned after durable anonymous input wiring"}
        )
    if payload.chart.family != "natal":
        raise _unsupported({"chart.family": "M2 supports natal astronomical snapshots only"})
    incompatible: dict[str, str] = {}
    if payload.settings.zodiac != "tropical":
        incompatible["settings.zodiac"] = "M2 supports tropical positions only"
    if payload.settings.center != "geocentric":
        incompatible["settings.center"] = "M2 supports geocentric positions only"
    if payload.settings.coordinate_frame == "horizontal":
        incompatible["settings.coordinate_frame"] = (
            "horizontal coordinates require M7 observer work"
        )
    if incompatible:
        raise _unsupported(incompatible)

    assert payload.subject.subject_version_id is not None
    try:
        subject_version = request.app.state.workflow_store.get_subject_version(
            payload.subject.subject_version_id
        )
    except WorkflowRecordNotFound as exc:
        raise ProblemException(
            status=status.HTTP_404_NOT_FOUND,
            code=ErrorCode.NOT_FOUND,
            detail="Subject version was not found.",
        ) from exc

    request_document = payload.model_dump(mode="json", exclude_none=True)
    try:
        snapshot = create_astronomical_snapshot(
            snapshot_id=f"calculation-{uuid4()}",
            request_payload=request_document,
            subject_version=subject_version,
            now=datetime.now(UTC),
            engine_version=request.app.state.settings.service_version,
            adapter=SwissEphemerisAdapter(
                moshier_fallback="record",
                ephemeris_path=request.app.state.settings.swiss_ephemeris_path,
            ),
        )
    except AstronomicalSnapshotInputError as exc:
        raise _unsupported({"subject.time_spec": str(exc)}) from exc
    request.app.state.workflow_store.put_snapshot(snapshot)
    return snapshot


@router.get("/calculations/{snapshot_id}/tables/{table_id}")
async def get_calculation_table(
    snapshot_id: str,
    table_id: str,
    request: Request,
    output_format: Literal["json", "csv"] = Query(default="json", alias="format"),
) -> Response:
    try:
        snapshot = request.app.state.workflow_store.get_snapshot(snapshot_id)
    except WorkflowRecordNotFound as exc:
        raise ProblemException(
            status=status.HTTP_404_NOT_FOUND,
            code=ErrorCode.NOT_FOUND,
            detail="Calculation snapshot was not found.",
        ) from exc
    try:
        table = build_snapshot_table(snapshot, table_id)
    except SnapshotTableError as exc:
        raise ProblemException(
            status=status.HTTP_409_CONFLICT,
            code=ErrorCode.INVALID_REQUEST,
            detail=str(exc),
        ) from exc

    headers = {
        "Content-Disposition": f'attachment; filename="{snapshot_id}-{table_id}.{output_format}"',
        "X-Interstellar-Snapshot-ID": snapshot_id,
        "X-Interstellar-Input-Fingerprint": snapshot["input_fingerprint"],
    }
    if output_format == "csv":
        return Response(
            content=table.to_csv().encode("utf-8"),
            media_type="text/csv; charset=utf-8",
            headers=headers,
        )
    return Response(
        content=table_json_bytes(table),
        media_type="application/json",
        headers=headers,
    )


@router.get("/calculations/{snapshot_id}/exports/natal-technical")
async def get_natal_technical_export(
    snapshot_id: str,
    request: Request,
    output_format: Literal["markdown", "plaintext"] = Query(
        default="markdown",
        alias="format",
    ),
) -> Response:
    try:
        snapshot = request.app.state.workflow_store.get_snapshot(snapshot_id)
    except WorkflowRecordNotFound as exc:
        raise ProblemException(
            status=status.HTTP_404_NOT_FOUND,
            code=ErrorCode.NOT_FOUND,
            detail="Calculation snapshot was not found.",
        ) from exc
    try:
        document = render_natal_technical_document(
            snapshot,
            output_format=output_format,
        )
    except NatalTechnicalExportError as exc:
        raise ProblemException(
            status=status.HTTP_409_CONFLICT,
            code=ErrorCode.INVALID_REQUEST,
            detail=str(exc),
        ) from exc
    extension = "md" if output_format == "markdown" else "txt"
    return Response(
        content=document.encode("utf-8"),
        media_type=(
            "text/markdown; charset=utf-8"
            if output_format == "markdown"
            else "text/plain; charset=utf-8"
        ),
        headers={
            "Content-Disposition": (
                f'attachment; filename="{snapshot_id}-natal-technical.{extension}"'
            ),
            "X-Interstellar-Snapshot-ID": snapshot_id,
            "X-Interstellar-Input-Fingerprint": snapshot["input_fingerprint"],
        },
    )