# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Interstellar Professional Astrology API
  version: 1.0.0-alpha.1
  description: >-
    Versioned deterministic western astrology API. All calculations resolve through an
    AnalysisRecipe, preserve dataset/rule provenance, and expose application/problem+json errors.
servers:
  - url: /api/v1
    description: Same-origin V1 API
security:
  - bearerAuth: []
...
paths:
...
  /calculations:
    get:
      tags: [Calculations]
      operationId: listCalculations
      parameters: [{$ref: '#/components/parameters/PageCursor'}, {$ref: '#/components/parameters/PageLimit'}]
      responses:
        '200':
          description: Calculation snapshot page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CalculationPage'}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Calculations]
      operationId: createCalculation
      security: [{}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/PreferAsync'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChartRequest'}
      responses:
        '201': {$ref: '#/components/responses/CalculationCreated'}
        '202': {$ref: '#/components/responses/JobAccepted'}
        default: {$ref: '#/components/responses/Problem'}
  /calculations/{id}:
    get:
      tags: [Calculations]
      operationId: getCalculation
      security: [{}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ResourceId'}, {$ref: '#/components/parameters/AnonymousAccessToken'}]
      responses:
        '200':
          description: Immutable calculation snapshot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CalculationSnapshot'}
        default: {$ref: '#/components/responses/Problem'}
  /calculations/{id}/tables/{table_id}:
    get:
      tags: [Calculations, Renders]
      operationId: getCalculationTable
      security: [{}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ResourceId'}
        - name: table_id
          in: path
          required: true
          schema: {type: string, minLength: 1, maxLength: 160}
        - name: format
          in: query
          required: false
          schema: {type: string, enum: [json, csv], default: json}
        - {$ref: '#/components/parameters/AnonymousAccessToken'}
      responses:
        '200':
          description: Deterministic table projection from the immutable snapshot
          headers:
            X-Interstellar-Snapshot-ID: {schema: {type: string}}
            X-Interstellar-Input-Fingerprint: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SnapshotTable'}
            text/csv:
              schema: {type: string}
        default: {$ref: '#/components/responses/Problem'}
  /calculations/{id}/exports/natal-technical:
    get:
      tags: [Calculations]
      operationId: exportNatalTechnicalDocument
      security: [{}, {bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ResourceId'}
        - name: format
          in: query
          required: false
          schema: {type: string, enum: [markdown, plaintext], default: markdown}
      responses:
        '200':
          description: Complete deterministic natal technical document
          headers:
            X-Interstellar-Snapshot-ID: {schema: {type: string}}
            X-Interstellar-Input-Fingerprint: {schema: {type: string}}
          content:
            text/markdown: {schema: {type: string}}
            text/plain: {schema: {type: string}}
        default: {$ref: '#/components/responses/Problem'}
  /calculations/{id}/interpretations/contextual:
    post:
      tags: [Interpretations]
      operationId: createNatalContextualInterpretations
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ContextualInterpretationBatchRequest'}
      responses:
        '200':
          description: Deterministic item-level interpretations; AI is never used
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ContextualInterpretationBatchResponse'}
        default: {$ref: '#/components/responses/Problem'}
  /ai/providers:
    get:
      tags: [AI Connectors]
      operationId: listNatalAiProviders
      security: [{}, {bearerAuth: []}]
      responses:
        '200':
          description: Optional downstream AI provider catalog and calculation boundary
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AiProviderCatalog'}
        default: {$ref: '#/components/responses/Problem'}
  /ai/analyses:
    post:
      tags: [AI Connectors]
      operationId: createNatalAiAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NatalAiAnalysisRequest'}
      responses:
        '202':
          description: Optional downstream AI analysis accepted when the provider is configured
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId: {name: id, in: path, required: true, schema: {type: string, minLength: 1}}
...
    PageCursor: {name: 'page[cursor]', in: query, required: false, schema: {type: string}}
    PageLimit: {name: 'page[limit]', in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 16, maxLength: 200}}
...
    PreferAsync: {name: Prefer, in: header, required: false, schema: {type: string, enum: [respond-async]}}
    AnonymousAccessToken: {name: access_token, in: query, required: false, description: Short-lived signed token for an anonymous calculation, schema: {type: string}}
  schemas:
...
    PageInfo: {$ref: '../packages/canonical-schema/common.schema.json#/$defs/PageInfo'}
...
    ChartRequest: {$ref: '../packages/canonical-schema/chart-request.schema.json'}
...
    CalculationSnapshot: {$ref: '../packages/canonical-schema/calculation-snapshot.schema.json'}
    SnapshotTable:
      type: object
      additionalProperties: false
      required: [table_id, metadata, columns, rows]
      properties:
        table_id: {type: string, minLength: 1, maxLength: 160}
        metadata:
          type: object
          additionalProperties: false
          required: [snapshot_id, input_fingerprint, engine_version]
          properties:
            snapshot_id: {type: string, minLength: 1}
            input_fingerprint: {type: string, minLength: 1}
            engine_version: {type: string, minLength: 1}
        columns: {type: array, minItems: 1, uniqueItems: true, items: {type: string}}
        rows: {type: array, items: {type: object, additionalProperties: true}}
    ContextualInterpretationItemRequest:
      type: object
      additionalProperties: false
      required: [item_kind, result_path, locale]
      properties:
        item_kind:
          type: string
          enum: [point_intrinsic, point_in_sign, point_in_house, motion, natal_aspect, house_cusp_ruler, structure_indicator, classical_condition]
        result_path: {type: string, pattern: '^/result/', minLength: 9, maxLength: 1000}
        locale: {type: string, enum: [zh-CN, en-US], default: zh-CN}
    ContextualInterpretationBatchRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 200
          items: {$ref: '#/components/schemas/ContextualInterpretationItemRequest'}
    ContextualInterpretationBatchResponse:
      type: object
      additionalProperties: false
      required: [schema_version, snapshot_id, interpretations, generation_mode, ai_used]
      properties:
        schema_version: {type: string}
        snapshot_id: {type: string}
        interpretations: {type: array, items: {type: object, additionalProperties: true}}
        generation_mode: {type: string, const: deterministic_rule_template}
        ai_used: {type: boolean, const: false}
    AiProviderCatalog:
      type: object
      additionalProperties: false
      required: [providers, calculation_boundary]
      properties:
        providers:
          type: array
          items:
            type: object
            additionalProperties: true
        calculation_boundary: {type: string}
    NatalAiAnalysisRequest:
      type: object
      additionalProperties: false
      required: [snapshot_id, provider_id, model_id, document_format, consent_to_send_snapshot, store_response]
      properties:
        snapshot_id: {type: string, minLength: 1, maxLength: 160}
        provider_id: {type: string, enum: [openai, moonshot]}
        model_id: {type: string, enum: [gpt, kimi]}
        document_format: {type: string, enum: [markdown, plaintext, json], default: markdown}
        analysis_focus: {type: [string, 'null'], maxLength: 4000}
        consent_to_send_snapshot: {type: boolean}
        store_response: {type: boolean, default: true}
...
    Job: {$ref: '../packages/canonical-schema/job.schema.json'}
    ProblemError: {$ref: '../packages/canonical-schema/problem-error.schema.json'}
...
    CalculationPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/CalculationSnapshot'}}
        page: {$ref: '#/components/schemas/PageInfo'}
...
  responses:
    Problem:
      description: RFC 9457-style problem response
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ProblemError'}
...
    CalculationCreated:
      description: Calculation completed within synchronous budget
      headers:
        X-RateLimit-Limit: {schema: {type: integer}}
        X-RateLimit-Remaining: {schema: {type: integer}}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CalculationSnapshot'}
    JobAccepted:
      description: Long-running operation accepted
      headers:
        Location: {schema: {type: string, format: uri-reference}}
        Retry-After: {schema: {type: integer, minimum: 1}}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Job'}
...